tools/product-impact-analyzer/requirements.txt
# Product Impact Analyzer Dependencies

# Google API clients
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
gspread>=5.11.0

# Google Ads API
google-ads>=23.0.0

# Data handling
pandas>=2.0.0

# AI
anthropic>=0.7.0

# Reporting
jinja2>=3.1.2
